backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic-settings==2.6.1
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
email-validator==2.1.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
slowapi==0.1.9
authlib==1.3.0
google-auth==2.34.0
requests==2.32.3
celery==5.3.4
redis==5.0.1
openai==1.12.0